i18n

+++ utils-support-common-starter/src/main/java/com/chua/common/support/database/transfer/SinkConfig.java
@@ -14,10 +14,10 @@ public class SinkConfig {
      * 数据开始位置
      */
     @Builder.Default
-    private long offset = 0;
+    private int offset = 0;
     /**
      * 限制数量
      */
     @Builder.Default
-    private long limit = 1000;
+    private int limit = 1000;
 }
